modified:   UNITE_Main.X/CommandParser.c 	modified:   UNITE_Main.X/CommandParser.h 	modified:   UNITE_Main.X/SampleManager.c 	modified:   UNITE_Main.X/SampleManager.h 	modified:   UNITE_Main.X/SystemConfiguration.c 	modified:   UNITE_Main.X/SystemConfiguration.h 	modified:   UNITE_Main.X/TransmitManager.c 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/SatelliteMode.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/main.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/mcc_generated_files/mcc.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/mcc_generated_files/tmr2.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/mcc_generated_files/tmr3.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/mcc_generated_files/tmr5.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/mcc_generated_files/uart2.o 	modified:   UNITE_Main.X/build/XC16_24FJ256GA110/production/mcc_generated_files/uart3.o 	modified:   UNITE_Main.X/dist/XC16_24FJ256GA110/production/memoryfile.xml 	modified:   UNITE_Main.X/mcc_generated_files/tmr5.c 	modified:   UNITE_Main.X/mcc_generated_files/uart3.c 	modified:   UNITE_Main.X/nbproject/private/private.xml

+++ UNITE_Main.X/SampleManager.c
@@ -59,16 +59,16 @@ uint16_t GPS_BUFFER_SIZE = 33;
   Instrument Timer Properties
  ******************************/
 
-int currentLangmuirProbeWait = 0;
-int currentMagnetometerWait = 0;
-int currentTemperatureWait = 0;
-int currentGPSWait = 0;
+unsigned long currentLangmuirProbeWait = 0;
+unsigned long currentMagnetometerWait = 0;
+unsigned long currentTemperatureWait = 0;
+unsigned long currentGPSWait = 0;
 
 int langmuirProbeCallbackCount = 0;
 int magnetometerCallbackCount = 0;
 
-int currentLangmuirProbeSweepProgress = 0;
-int currentMagnetometerSweepProgress = 0;
+unsigned long currentLangmuirProbeSweepProgress = 0;
+unsigned long currentMagnetometerSweepProgress = 0;
 
 int currentMagnetometerOrbitProgress = 0;
